backend: extract chat and user routers
